conf

+++ Continuous delivery for Intune/HKCU/Install-HKCU.ps1
@@ -1,22 +1,27 @@
 function Install-HKCU {
     Param(
         $URL,
-        $Detection
+        $detection
     )
-    $TempRegFile = $env:TEMP + "\Temp.reg"
-    Remove-Item $TempRegFile -Force | Out-Null
-    Invoke-WebRequest -Uri $URL -OutFile $TempRegFile
-    $regfile = Get-Content $TempRegFile
-    $hives = Get-ChildItem -Path REGISTRY::HKEY_USERS | Select-Object -ExpandProperty Name
-    foreach ($hive in $hives) {
-        if (!($hive -like "*_Classes")) {
-            $newregfile = $regfile -replace "HKEY_CURRENT_USER",$hive
-            Set-Content -Path $TempRegFile -Value $newregfile
-            $Arguments = "/s $TempRegFile"
-            Start-Process "regedit.exe" -ArgumentList $Arguments -Wait
+
+    $runDetectionRule = Invoke-Expression -Command $detection
+    
+    If (!($runDetectionRule -eq $true)) {
+        $TempRegFile = $env:TEMP + "\Temp.reg"
+        Remove-Item $TempRegFile -Force | Out-Null
+        Invoke-WebRequest -Uri $URL -OutFile $TempRegFile
+        $regfile = Get-Content $TempRegFile
+        $hives = Get-ChildItem -Path REGISTRY::HKEY_USERS | Select-Object -ExpandProperty Name
+        foreach ($hive in $hives) {
+            if (!($hive -like "*_Classes")) {
+                $newregfile = $regfile -replace "HKEY_CURRENT_USER",$hive
+                Set-Content -Path $TempRegFile -Value $newregfile
+                $Arguments = "/s $TempRegFile"
+                Start-Process "regedit.exe" -ArgumentList $Arguments -Wait
+            }
         }
+        Remove-Item $TempRegFile -Force
     }
-    Remove-Item $TempRegFile -Force
 }
 
 $RegFileConf = $env:TEMP + "\RegFileConfig.JSON"
